PR comments refactor to make the splits a little nicer
